TextField Controller implemented
In order for us to be able to submit a search string, we first need to be able to retrieve the contents of the TextField at the time of searching, and now we can! Just a proof-of-concept as of right now, I've got to make some tweaks to retrieveCardDatabase.dart first so that i can grab the entirety of AtomicCards.json locally to prove you can actually search the .json using SearchCardDatabase. After that I'll look into implementing FirebaseDB as a backend to store AtomicCards.json to free up space on the phone locally.