Initial FirbaseDB Link Implemented
Set up a Firebase project under nathaniel.dwyer.unh Gmail account. Linked the Firebase project to both the Android and iOS versions of the Grimoire app. Made a realtime database for the project & added basic functionality to the main.dart file to write to the FirebaseDB to make sure a connection was established.