Fixing FirebaseDB cardName keys
Apparently FirebaseDB doesn't like certain special symbols in the key values of the database: [/#&.\]. This is causing some cardName keys to be invalid (For Example: 'Richard Garfield, Ph.D.' and 'Wear // Tear'), amounting to a total of 315 illegal keys. To accommodate this I used a str.replaceAll() to replace any special character in cardName with a dash '-'.