Rudimentary cardData Retrieval from FirebaseDB
implemented a basic data retrieval method  - for the time being it just takes the string value input by the user and searches for an exact match in the FirebaseDB and returns the data on that card to the phone screen. the user can continue to search using different string values in the textField and searching using the 'Search FirebaseDB' button.